バリデーションチェック処理を共通化
